lwgeom_scale: recompute bounding box if needed
git-svn-id: http://svn.osgeo.org/postgis/trunk@13653 b70326c6-7e19-0410-871a-916f4a2858ee

+++ liblwgeom/cunit/cu_tester.h
@@ -25,8 +25,8 @@ typedef void (*PG_SuiteSetup)(void);
 
 #define ASSERT_DOUBLE_EQUAL(o,e) do { \
   if ( o != e ) \
-    fprintf(stderr, "[%s:%d]\n Expected: %g\n Obtained: %g\n", __FILE__, __LINE__, (e), (o)); \
-  CU_ASSERT_EQUAL(o,e); \
+    fprintf(stderr, "[%s:%d]\n Expected: %g\n Obtained: %g\n", __FILE__, __LINE__, (double)(e), (o)); \
+  CU_ASSERT_EQUAL(o,(double)e); \
 } while (0);
 
 #define ASSERT_INT_EQUAL(o,e) do { \
